Multipage site (pages/..), upd admin_panel, upd movie_suggestion, upd logs data_proc
